Updated index.html; added notes!
Added a lot of notes for improvement.
Replaced the select field with an input field so users can look for more than just bars and pharmacies.
I hadn't realized the number of coding mistakes that were made in this file. (No title element, and all the script declarations are in the body.  That needs to be fixed!)