Moved around a few things.
Ugh! Github's editor is so hard to look at!  I'm going to see if I can get this who "pull/push" thing figured out and edit this on my computer.

Added header and footer sections.
Put most of the external javascript stuff in the header where it belongs.
We should consider putting our functions in a JavaScript class (which is really an Object) called Dive then putting that stuff into a file called Dive.js for others to use.