Update index.html (Read the commit comments)
Out is Jeremia's bad HTML code. (Scripts go in the header!)
In is the Dive.js file, which contains a Dive class to make Dive objects that get a list of bus routes based on some external HTML file.
To insure that that other file doesn't disappear like the Transit API file that was hosted on Heroku, and is now causing things not to work, I'll need to localize all the files.  (As impressive as that CDN is you have your data put on, it's better that we put all of our eggs--or files--in one basket--or website.  We can worry about impressing some VCs in SV with our CDN sometime in the future where people have an acronym to describe "later".)